refactor(backend): unify internal api key
