+++ apps/backend/src/main/java/com/sos/backend/global/internal/InternalApiProperties.java
@@ -1,4 +1,4 @@
-package com.sos.backend.global.common.config;
+package com.sos.backend.global.internal;
 
 import jakarta.validation.constraints.NotBlank;
 import org.springframework.boot.context.properties.ConfigurationProperties;
@@ -11,7 +11,7 @@ import org.springframework.validation.annotation.Validated;
 @ConfigurationProperties(prefix = "app.internal")
 @Validated
 public record InternalApiProperties(
-    @NotBlank
-    String apiKey
+    @NotBlank String apiKey,
+    @NotBlank String baseUrl
 ) {
 }
