+++ apps/backend/src/main/resources/application-local.yml
@@ -31,6 +31,12 @@ spring:
 server:
   port: 8080
 
+app:
+  internal:
+    # game-engine >> backend 서버 간 호출 인증용 공유 시크릿. game-engine 의 INTERNAL_API_KEY 와 동일해야 함.
+    api-key: ${INTERNAL_API_KEY:local-internal-api-key-at-least-16-bytes}
+    base-url: ${GAME_ENGINE_BASE_URL:http://localhost:8000}
+
 jwt:
   secret: ${JWT_SECRET:local-jwt-secret-key-at-least-32-bytes-123}
   access-token-expiration: 1800000
@@ -41,6 +47,3 @@ auth:
     secret: ${VERIFICATION_TOKEN_SECRET:local-verification-token-secret-at-least-32b}
     expires-in-seconds: 300
 
-game-engine:
-  base-url: ${GAME_ENGINE_BASE_URL:http://localhost:8000}
-  internal-api-key: ${INTERNAL_API_KEY:ci-smoke-default-internal-key-please-override-min-32-bytes-string}
